LOG信息: 1. APP端 -- 修改收到传阅详情接口, 增加一个 接收人总数的值. 2. PC端 -- 修改 最近联系人 / 同部门 / 搜索 接口 3. APP端 -- 修改 传阅详情 接口 4. APP/PC端 -- 修改 删除已发传阅 接口 5. PC端 -- 修改 已收传阅列表, OA联系人 接口 6. PC端 -- 修改 传阅详情添加传阅对象的 接口 7.【发起传阅】收到传阅添加人员和收件人的搜索按钮、右侧传阅对象，可以加组织 8.【佰钧成】【已发/收到的传阅】打开传阅时，会提示调研工作 9. PC端 -- 修改 传阅详情删除传阅对象 / 已发传阅列表的 接口 10. PC端 -- 根据mailId单独查询传阅内容返回给前端 11. PC端 -- 修改 已发传阅列表的 接口

+++ src/main/java/com/sunnsoft/sloa/actions/web/received/BatchDeleteReceived.java
@@ -37,6 +37,15 @@ public class BatchDeleteReceived extends BaseParameter {
 
 		// 根据传阅ID查询
 		Mail mail = Services.getMailService().createHelper().getMailId().Eq(mailId).uniqueResult();
+		List<Receive> ress = mail.getReceives();
+		if (ress .size() == 1) {
+			success = false;
+			code = "404";
+			msg = "传阅对象只有一个用户时, 不允许删除";
+			json = "null";
+			return Results.GLOBAL_FORM_JSON;
+		}
+		Receive lastReceive = ress.get(ress.size() - 1);
 		try {
 
 			// 遍历收件人数组
@@ -49,8 +58,13 @@ public class BatchDeleteReceived extends BaseParameter {
 				// 获取要删除的传阅是谁添加的
 				Long reDifferentiate = receive.getReDifferentiate();
 
+				if (lastReceive.getUserId() == receive.getUserId()){
+					continue;
+				}
+
 				// 判断 删除传阅对象, 有两种情况 : 一. 是发件人可以删除所有的传阅对象 , 二是 谁添加该传阅对象 谁才有权限删除
 				if (reDifferentiate.equals(userId) || mail.getUserId() == userId) {
+
 					// 执行删除
 					Services.getReceiveService().delete(receive);
